Revert "LPD-95804 move text extraction logic from service to api"
This reverts commit c3d31f8c752b1edcc019fecf2e290176d4c2ce4d.
