--- modules/apps/document-library/document-library-api/src/main/java/com/liferay/document/library/text/DLFileEntryTextProvider.java
@@ -1,17 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.document.library.text;
-
-import com.liferay.document.library.kernel.model.DLFileEntry;
-
-/**
- * @author Jhosseph Gonzalez
- */
-public interface DLFileEntryTextProvider {
-
-	public String getText(DLFileEntry dlFileEntry);
-
-}
