--- modules/apps/document-library/document-library-api/src/main/resources/com/liferay/document/library/text/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
