+++ modules/apps/document-library/document-library-service/src/main/java/com/liferay/document/library/internal/search/spi/model/index/contributor/DLFileEntryModelDocumentContributor.java
@@ -5,12 +5,18 @@
 
 package com.liferay.document.library.internal.search.spi.model.index.contributor;
 
+import com.liferay.change.tracking.model.CTCollection;
+import com.liferay.change.tracking.service.CTCollectionLocalService;
+import com.liferay.document.library.internal.configuration.DLIndexerConfiguration;
+import com.liferay.document.library.kernel.exception.NoSuchFileException;
 import com.liferay.document.library.kernel.model.DLFileEntry;
 import com.liferay.document.library.kernel.model.DLFileEntryMetadata;
 import com.liferay.document.library.kernel.model.DLFileEntryMetadataTable;
 import com.liferay.document.library.kernel.model.DLFileVersion;
 import com.liferay.document.library.kernel.service.DLFileEntryMetadataLocalService;
-import com.liferay.document.library.text.DLFileEntryTextProvider;
+import com.liferay.document.library.kernel.store.DLStore;
+import com.liferay.document.library.kernel.store.DLStoreRequest;
+import com.liferay.document.library.security.io.InputStreamSanitizer;
 import com.liferay.dynamic.data.mapping.model.DDMStructure;
 import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
 import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
@@ -19,11 +25,15 @@ import com.liferay.dynamic.data.mapping.util.DDMFormValuesIndexer;
 import com.liferay.dynamic.data.mapping.util.DDMIndexer;
 import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerList;
 import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerListFactory;
+import com.liferay.petra.io.StreamUtil;
 import com.liferay.petra.sql.dsl.DSLQueryFactoryUtil;
 import com.liferay.petra.sql.dsl.query.DSLQuery;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
+import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
+import com.liferay.portal.kernel.change.tracking.CTCollectionThreadLocal;
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -33,15 +43,24 @@ import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.kernel.search.ReindexCacheThreadLocal;
 import com.liferay.portal.kernel.search.RelatedEntryIndexer;
 import com.liferay.portal.kernel.search.RelatedEntryIndexerRegistry;
+import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.PrefsProps;
+import com.liferay.portal.kernel.util.PropsKeys;
+import com.liferay.portal.kernel.util.PropsUtil;
 import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.TextExtractor;
 import com.liferay.portal.repository.liferayrepository.model.LiferayFileEntry;
 import com.liferay.portal.search.ml.embedding.text.TextEmbeddingDocumentContributor;
 import com.liferay.portal.search.spi.model.index.contributor.ModelDocumentContributor;
 import com.liferay.trash.TrashHelper;
+
+import java.io.IOException;
+
+import java.nio.charset.StandardCharsets;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -53,12 +72,14 @@ import org.osgi.framework.BundleContext;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.service.component.annotations.Modified;
 import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Michael C. Han
  */
 @Component(
+	configurationPid = "com.liferay.document.library.internal.configuration.DLIndexerConfiguration",
 	property = "indexer.class.name=com.liferay.document.library.kernel.model.DLFileEntry",
 	service = ModelDocumentContributor.class
 )
@@ -79,7 +100,7 @@ public class DLFileEntryModelDocumentContributor
 
 			_addFile(
 				document, Field.getLocalizedName(defaultLocale, Field.CONTENT),
-				dlFileEntry);
+				dlFileEntry, dlFileVersion);
 
 			document.addKeyword(
 				Field.CLASS_TYPE_ID, dlFileEntry.getFileEntryTypeId());
@@ -165,7 +186,11 @@ public class DLFileEntryModelDocumentContributor
 	}
 
 	@Activate
-	protected void activate(BundleContext bundleContext) {
+	protected void activate(
+		BundleContext bundleContext, Map<String, Object> properties) {
+
+		modified(properties);
+
 		_serviceTrackerList = ServiceTrackerListFactory.open(
 			bundleContext, DDMFormValuesIndexer.class);
 	}
@@ -175,22 +200,128 @@ public class DLFileEntryModelDocumentContributor
 		_serviceTrackerList.close();
 	}
 
+	@Modified
+	protected void modified(Map<String, Object> properties) {
+		_dlIndexerConfiguration = ConfigurableUtil.createConfigurable(
+			DLIndexerConfiguration.class, properties);
+	}
+
 	private void _addFile(
-		Document document, String fieldName, DLFileEntry dlFileEntry) {
+		Document document, String fieldName, DLFileEntry dlFileEntry,
+		DLFileVersion dlFileVersion) {
 
-		String text = _dlFileEntryTextProvider.getText(dlFileEntry);
-
-		if (text == null) {
+		if (!_isIndexContent(dlFileEntry)) {
 			return;
 		}
 
-		document.addText(fieldName, text);
+		try {
+			String text = _extractText(dlFileEntry, dlFileVersion);
 
-		_textEmbeddingDocumentContributor.contribute(
-			document, dlFileEntry,
-			StringBundler.concat(
-				dlFileEntry.getTitle(), StringPool.PERIOD, StringPool.SPACE,
-				text));
+			if (text != null) {
+				document.addText(fieldName, text);
+
+				_textEmbeddingDocumentContributor.contribute(
+					document, dlFileEntry,
+					StringBundler.concat(
+						dlFileEntry.getTitle(), StringPool.PERIOD,
+						StringPool.SPACE, text));
+			}
+		}
+		catch (Throwable throwable) {
+			if (_log.isWarnEnabled()) {
+				_log.warn(
+					"Unable to extract text from file version " +
+						dlFileVersion.getFileVersionId());
+			}
+
+			if (_log.isDebugEnabled()) {
+				_log.debug(throwable);
+			}
+		}
+	}
+
+	private String _extractText(
+			DLFileEntry dlFileEntry, DLFileVersion dlFileVersion)
+		throws IOException, PortalException {
+
+		int dlFileIndexingMaxSize = GetterUtil.getInteger(
+			PropsUtil.get(PropsKeys.DL_FILE_INDEXING_MAX_SIZE));
+
+		String indexVersionLabel = dlFileVersion.getStoreFileName() + ".index";
+
+		if (_dlIndexerConfiguration.cacheTextExtraction()) {
+			try {
+				String string = StreamUtil.toString(
+					_dlStore.getFileAsStream(
+						dlFileEntry.getCompanyId(),
+						dlFileEntry.getDataRepositoryId(),
+						dlFileEntry.getName(), indexVersionLabel));
+
+				if (string.length() <= dlFileIndexingMaxSize) {
+					if (string.isEmpty()) {
+						return null;
+					}
+
+					return string;
+				}
+
+				_dlStore.deleteFile(
+					dlFileEntry.getCompanyId(),
+					dlFileEntry.getDataRepositoryId(), dlFileEntry.getName(),
+					indexVersionLabel);
+			}
+			catch (NoSuchFileException noSuchFileException) {
+				if (_log.isDebugEnabled()) {
+					_log.debug(
+						"Unable to get cached text extraction" +
+							noSuchFileException);
+				}
+			}
+		}
+
+		String text = null;
+
+		try {
+			text = _textExtractor.extractText(
+				_inputStreamSanitizer.sanitize(
+					dlFileVersion.getContentStream(false)),
+				dlFileIndexingMaxSize);
+		}
+		catch (PortalException portalException) {
+			if (_log.isDebugEnabled()) {
+				_log.debug("Unable to get input stream", portalException);
+			}
+		}
+
+		if (_dlIndexerConfiguration.cacheTextExtraction() &&
+			!_isReadOnlyCtCollection()) {
+
+			byte[] bytes = null;
+
+			if (text == null) {
+				bytes = new byte[0];
+			}
+			else {
+				bytes = text.getBytes(StandardCharsets.UTF_8);
+			}
+
+			_dlStore.addFile(
+				DLStoreRequest.builder(
+					dlFileEntry.getCompanyId(),
+					dlFileEntry.getDataRepositoryId(), dlFileEntry.getName()
+				).className(
+					dlFileEntry.getModelClassName()
+				).classPK(
+					dlFileEntry.getFileEntryId()
+				).sourceFileName(
+					dlFileEntry.getFileName()
+				).versionLabel(
+					indexVersionLabel
+				).build(),
+				bytes);
+		}
+
+		return text;
 	}
 
 	private Map<DDMStructure, DDMFormValues> _getDDMFormValues(
@@ -306,6 +437,25 @@ public class DLFileEntryModelDocumentContributor
 			dlFileVersionId, Collections.emptyMap());
 	}
 
+	private boolean _isIndexContent(DLFileEntry dlFileEntry) {
+		String[] ignoreExtensions = _prefsProps.getStringArray(
+			PropsKeys.DL_FILE_INDEXING_IGNORE_EXTENSIONS, StringPool.COMMA);
+
+		return !ArrayUtil.contains(
+			ignoreExtensions, StringPool.PERIOD + dlFileEntry.getExtension());
+	}
+
+	private boolean _isReadOnlyCtCollection() throws PortalException {
+		if (CTCollectionThreadLocal.isProductionMode()) {
+			return false;
+		}
+
+		CTCollection ctCollection = _ctCollectionLocalService.getCTCollection(
+			CTCollectionThreadLocal.getCTCollectionId());
+
+		return ctCollection.isReadOnly();
+	}
+
 	private void _processDDMIndexer(Document document, long dlFileVersionId) {
 		Map<DDMStructure, DDMFormValues> ddmFormValuesMap = _getDDMFormValues(
 			dlFileVersionId);
@@ -347,6 +497,9 @@ public class DLFileEntryModelDocumentContributor
 		DLFileEntryModelDocumentContributor.class);
 
 	@Reference
+	private CTCollectionLocalService _ctCollectionLocalService;
+
+	@Reference
 	private DDMIndexer _ddmIndexer;
 
 	@Reference
@@ -358,11 +511,19 @@ public class DLFileEntryModelDocumentContributor
 	@Reference
 	private DLFileEntryMetadataLocalService _dlFileEntryMetadataLocalService;
 
+	private volatile DLIndexerConfiguration _dlIndexerConfiguration;
+
 	@Reference
-	private DLFileEntryTextProvider _dlFileEntryTextProvider;
+	private DLStore _dlStore;
+
+	@Reference
+	private InputStreamSanitizer _inputStreamSanitizer;
 
 	@Reference
 	private Portal _portal;
+
+	@Reference
+	private PrefsProps _prefsProps;
 
 	@Reference
 	private RelatedEntryIndexerRegistry _relatedEntryIndexerRegistry;
@@ -371,6 +532,9 @@ public class DLFileEntryModelDocumentContributor
 
 	@Reference
 	private TextEmbeddingDocumentContributor _textEmbeddingDocumentContributor;
+
+	@Reference
+	private TextExtractor _textExtractor;
 
 	@Reference
 	private TrashHelper _trashHelper;
